+++ server/routes/comment.routes.js
@@ -1,13 +1,13 @@
 const express = require('express')
-// const auth = require('../middleware/auth.middleware')
+const auth = require('../middleware/auth.middleware')
 const chalk = require('chalk')
 const Comment = require('../models/Comment')
+const User = require('../models/User')
 const router = express.Router({mergeParams: true})
 
 // /api/comment
 router
   .route('/')
-  // +auth
   .get(async (req, res) => {
     try {
       const { orderBy, equalTo } = req.query
@@ -20,40 +20,37 @@ router
         errors: errors.array()
       })
     }
-  })
-  // +auth, 
+  }) 
   .post(async (req, res) => {
     try {
       const newComment = await Comment.create({
         ...req.body
-        // userId: req.user._id
       })
       res.status(201).send(newComment)
     } catch (e) {
       console.log(chalk.red('error'), e)
       res.status(500).json({
         message: 'На сервере проихошла ошибка, попробуйте позже.',
-        // errors: errors.array()
       })
     }
   })
-// +auth,
-router.delete('/:commentId', async (req, res) => {
+
+router.delete('/:commentId', auth, async (req, res) => {
   try {
     const { commentId } = req.params
     const removedComment = await Comment.findById(commentId)
+    const authedUser = await User.findById(req.user._id)
 
-    // if (removedComment.userId.toString() === req.user._id) {
+    if (removedComment.userId.toString() === req.user._id || authedUser.admin) {
       await removedComment.deleteOne()
       return res.send(null)
-    // } else {
-    //   return res.status(401).json({message: 'Unauthorized'})
-    // }
+    } else {
+      return res.status(401).json({message: 'Unauthorized'})
+    }
   } catch (e) {
     console.log(chalk.red('error'), e)
     res.status(500).json({
       message: 'На сервере проихошла ошибка, попробуйте позже.'
-      // errors: errors.array()
     })
   }
 })
